Make `userFromRequest` return a `User`
Hide the `Auth0User` type inside the client so the rest of VxDesign can
think in terms of the `User` type. This also allows us to remove the
`org()` method from the client, shrinking its surface area, which will
make it easier to mock.
